fix compile errors from filter and support_interator_tag (#877)
