+++ lib/utils/include/utils/type_traits.h
@@ -113,16 +113,6 @@ struct elements_satisfy<Cond, std::tuple<>> : std::true_type {};
 static_assert(
     elements_satisfy<is_equal_comparable, std::tuple<int, float>>::value, "");
 
-template <typename C, typename Tag>
-struct supports_iterator_tag
-    : std::is_base_of<Tag,
-                      typename std::iterator_traits<
-                          typename C::iterator>::iterator_category> {};
-
-#define CHECK_SUPPORTS_ITERATOR_TAG(TAG, ...)                                  \
-  static_assert(supports_iterator_tag<__VA_ARGS__, TAG>::value,                \
-                #__VA_ARGS__ " does not support required iterator tag " #TAG);
-
 template <typename T>
 using is_default_constructible = std::is_default_constructible<T>;
